+++ R/provider-openrouter.R
@@ -15,7 +15,7 @@ NULL
 #' @family chatbots
 #' @param api_key `r lifecycle::badge("deprecated")` Use `credentials` instead.
 #' @param credentials `r api_key_param("OPENROUTER_API_KEY")`
-#' @param model `r param_model("gpt-5.4")`
+#' @param model `r param_model("gpt-5.6-terra")`
 #' @param params Common model parameters, usually created by [params()].
 #' @inheritParams chat_openai
 #' @inherit chat_openai return
@@ -34,7 +34,7 @@ chat_openrouter <- function(
   echo = c("none", "output", "all"),
   api_headers = character()
 ) {
-  model <- set_default(model, "gpt-5.4")
+  model <- set_default(model, "gpt-5.6-terra")
   echo <- check_echo(echo)
 
   credentials <- as_credentials(