+++ R/provider-posit.R
@@ -23,7 +23,7 @@ NULL
 #'   credentials to use in place of the default OAuth device flow, either as a
 #'   named list of headers or as a function that modifies the request. You
 #'   should not usually need to set this.
-#' @param model `r param_model("claude-sonnet-4-6", "posit")`
+#' @param model `r param_model("claude-sonnet-5", "posit")`
 #' @param cache How long to cache inputs? Defaults to "5m" (five minutes).
 #'   Set to "none" to disable caching or "1h" to cache for one hour. This is
 #'   only supported for Claude models and is ignored for other models.
@@ -49,7 +49,7 @@ chat_posit <- function(
   echo = NULL
 ) {
   check_string(base_url)
-  model <- set_default(model, "claude-sonnet-4-6")
+  model <- set_default(model, "claude-sonnet-5")
   cache <- arg_match(cache)
   echo <- check_echo(echo)